rename layout to MainLayout
